Enforce that the tools are not run in the top-level directory

+++ fedora-build-all.py
@@ -66,6 +66,11 @@ def _do_newest_filtering(filelist):
 
 def main():
 
+    # check we're not top level
+    if os.path.exists('./application.py'):
+        print 'You cannot run these tools from the top level directory'
+        sys.exit(1)
+
     # remove appstream
     if os.path.exists('./appstream'):
         shutil.rmtree('./appstream')
